+++ migrations/sqlite/20260816090300_lifecycle.sql
@@ -1,5 +1,0 @@
--- +goose Up
-ALTER TABLE credbound_workspaces ADD COLUMN disabled_at DATETIME;
-
--- +goose Down
-ALTER TABLE credbound_workspaces DROP COLUMN disabled_at;